Split CLI P2P server setup
